python/02-use-cases/data_analysis_with_datalake/pyproject.toml
[project]
name = "data-analysis-with-datalake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentkit-sdk-python==0.4.3",
    "google-adk==1.21.0",
    "veadk-python==0.5.18",
    "python-dotenv==1.2.1",
    "lancedb==0.25.3",
    "volcengine-python-sdk[ark]>=5.0.1",
    "pyarrow==21.0.0",
    "duckdb==1.4.3",
    "pandas",
]
